requirements.txt
captcha
Django==3.1.5
mysqlclient==2.0.2
Pillow==8.0.1
pip==21.0.1
django-tinymce4-lite==1.8.0